benchmarking/scenarios/scenario1/requirements.txt
mysql-python
progressbar2
